libs/computer-server/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "cua-computer-server"
version = "0.1.0"
description = "Server component for the Computer-Use Interface (CUI) framework powering Cua"
authors = [
    { name = "TryCua", email = "gh@trycua.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pyautogui>=0.9.54",
    "pillow>=10.2.0",
    "aiohttp>=3.9.1",
    "mac-pyxa"
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Cocoa>=10.1",
    "pyobjc-framework-Quartz>=10.1",
    "pyobjc-framework-ApplicationServices>=10.1"
]
linux = [
    "python-xlib>=0.33"
]

[project.urls]
homepage = "https://github.com/trycua/cua"
repository = "https://github.com/trycua/cua"

[project.scripts]
cua-computer-server = "computer_server:run_cli"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["computer_server"]
package-data = {"computer_server" = ["py.typed"]}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0"
]
format = [
    "black>=23.0.0",
    "isort>=5.12.0"
]
dev = [
    "ruff>=0.0.241",
    "mypy>=0.971"
]

[tool.pdm.scripts]
api = "python -m computer_server"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I"]
fix = true

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false